chore: complete toveco to cardinal_vote rename across entire codebase
Systematically renamed all "toveco" references to "cardinal_vote" throughout
the codebase to achieve internal consistency with the GitHub repository name.

## Key Changes

### Module Structure
- Renamed main module: src/toveco_voting/ → src/cardinal_vote/
- Updated all import statements and module references

### Configuration & Settings
- Updated LOGO_PREFIX: "toveco" → "cardinal_vote" in config.py
- Modified all logo validation logic to use dynamic prefix
- Updated pyproject.toml package and script references

### Asset Files
- Renamed all logo files: toveco*.png → cardinal_vote*.png (11 files)
- Updated archive filenames: toveco-voting*.tar → cardinal-vote*.tar

### Infrastructure
- Updated Docker configuration and entrypoint scripts
- Modified GitHub Actions CI workflow imports
- Updated mypy configuration for new module name
- Updated CODEOWNERS file paths
- Fixed pre-commit hooks for new logo naming convention

### Documentation
- Updated README, DEVELOPMENT, and CLAUDE documentation
- Modified example filenames and directory structures
- Updated API documentation references

## Technical Fixes
- Fixed import ordering in alembic/env.py for ruff compliance
- Added missing type stubs for mypy (types-python-jose, types-passlib)
- Updated logo format validation in pre-commit hooks
- Added proper type ignore comments for untyped/missing library imports
- Added nosec comments for intentional security validation strings

## Validation
- ✅ Config module imports successfully
- ✅ LOGO_PREFIX confirmed as "cardinal_vote"
- ✅ All 11 logo files detected and functional
- ✅ Core application functionality preserved
- ✅ All pre-commit hooks passing

This completes the systematic transformation from ToVéCo-specific branding
to generic voting platform naming while maintaining full functionality.
