+++ docs/audit/affected-files.txt
@@ -1,0 +1,45 @@
+./alembic/env.py
+./API_DOCUMENTATION.md
+./API.md
+./CLAUDE.md
+./.claude/settings.local.json
+./deploy-from-tar.sh
+./DEPLOYMENT.md
+./DEPLOYMENT-TAR.md
+./deploy-server.sh
+./DEVELOPMENT.md
+./docker-compose.override.yml.example
+./docker-compose.production.yml
+./docker-compose.prod.yml
+./docker-compose.standalone.yml
+./docker-compose.yml
+./docker-entrypoint.sh
+./Dockerfile
+./docs/audit/pre-rename-toveco-references.txt
+./docs/brainstorming/2025-09-08-toveco-to-cardinal-vote-rename.md
+./docs/prps/cardinal-vote-rename.md
+./docs/prps/generalized-voting-platform.md
+./docs/tasks/cardinal-vote-rename.md
+./.env.example
+./.github/CODEOWNERS
+./.github/workflows/ci.yml
+./.github/workflows/release.yml
+./mypy.ini
+./node_modules/.cache/prettier/.prettier-caches/b70c50ca592e0d108ad6393036b30747ad295215.json
+./.pre-commit-config.yaml
+./PRP.md
+./pyproject.toml
+./README.md
+./scripts/deploy.sh
+./scripts/manage.sh
+./scripts/run.sh
+./src/toveco_voting/admin_manager.py
+./src/toveco_voting/config.py
+./src/toveco_voting/main.py
+./src/toveco_voting/models.py
+./static/app.js
+./templates/admin/logos.html
+./TEST_SUITE_SUMMARY.md
+./test_vote_details_fix.py
+./uv.lock
+./WORKFLOW-GENERALIZED.md
