+++ logos/cardinal_vote4.png
