+++ logos/cardinal_vote7.png
