+++ mypy.ini
@@ -14,7 +14,7 @@ warn_no_return = True
 warn_unreachable = True
 strict_equality = True
 
-[mypy-toveco_voting.models]
+[mypy-cardinal_vote.models]
 # SQLAlchemy models have complex typing that's handled by the plugin
 warn_return_any = False
 disallow_untyped_defs = False
