+++ src/cardinal_vote/admin_routes_simple.py
