+++ src/cardinal_vote/database_manager.py
